Split workspaces and move source files
Fixes #26
